Defer processing of styles to later point in the process
Processing has moved from ObflParser to FormatterCoreImpl. The idea now
is to move the evaluation of expressions from BlockContextManager to
FormatterCoreImpl so that styles can be applied on evaluated
expressions. Because TextSegment has a TextAttribute field now it is
even possible to apply the styles within BrailleTranslators (as is
already done with header and footer fields), with the benefit of having
possibly useful style information available during line
breaking. However this comes with the price of either losing the ability
to retain marker/anchor positions in between text segments, or losing
the ability to process several text segments together.

see issue https://github.com/joeha480/dotify/issues/118