Various changes to <evaluate/> (#99)
* Allow evaluation of "$volume" in volume transitions and in normal flow

* Allow <evaluate expression="$page"/>

i.e. don't require the variable to be wrapped in a "(round ...)" or "(format {0} ...)"

Better would be to completely remove ExpressionTools and resolve variables in ExpressionImpl.

* Don't collect id-to-volume mapping until pages have come out of SheetDataSource

Initially, for pages in the volume body, the id-to-volume mapping was
done in VolumeProvider, but since the previous commit it was done also
in PageSequenceBuilder#nextPage, so in two places.

For pages in pre-/post-content, the id-to-volume mapping was only done
in PageSequenceBuilder#nextPage.

This commit moves the logic to one place, in VolumeProvider. Apart
from simplifying the code, this also fixes an issue in the regression
test suite. I have no proper explanation for it.

* In VolumeProvider align nextBodyContents and updateVolumeContents methods more

* Be more strict about what is allowed in an expression

For example, don't allow "$foo" when the variable "foo" is not declared.

* Don't normalize space in evaluate expressions